If Judy is missing, use POSIX tree functions instead of linked list
On GNU/Linux (glibc) on my system, <search.h> is about 2.5x slower than
Judy. This is much much faster than a linked list.

The good news is, Miredo is now reasonably fast even without Judy. The
not so good news is, Miredo requires a GNU extension, tdestroy(), if
Judy is disabled.